# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/collaborators/{collaboratorId}/arrears:
    get:
      operationId: listCollaboratorArrears
      tags: [admin-finance]
      parameters:
        - in: path
          name: collaboratorId
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Confidential arrears and transfer-request records for a collaborator
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArrearsList" }
        "400": { description: Invalid collaborator id }
        "403": { description: Confidential finance data requires the ADMIN role }
        "404": { description: Collaborator not found }
    post:
      operationId: createCollaboratorArrear
      tags: [admin-finance]
      parameters:
        - in: path
          name: collaboratorId
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArrearInput" }
      responses:
        "201":
          description: Arrear created and the recipient granted own-finance visibility
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArrearResult" }
        "400": { description: Invalid arrear or recipient }
        "403": { description: Confidential finance data requires the ADMIN role }
  /admin/collaborators/{collaboratorId}/salary-records:
    get:
      operationId: listCollaboratorSalaryRecords
      tags: [admin-finance]
      parameters:
        - in: path
          name: collaboratorId
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Confidential salary records for a collaborator
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalaryRecordsList" }
        "400": { description: Invalid collaborator id }
        "403": { description: Confidential finance data requires the ADMIN role }
        "404": { description: Collaborator not found }
    post:
      operationId: assignCollaboratorSalaryRecord
      tags: [admin-finance]
      parameters:
        - in: path
          name: collaboratorId
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SalaryRecordInput" }
      responses:
        "201":
          description: Salary record assigned and own-finance visibility granted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalaryRecordResult" }
        "400": { description: Invalid record or non-collaborator recipient }
        "403": { description: Confidential finance data requires the ADMIN role }
        "409": { description: A record for this period already exists }
  /admin/salary-records/{id}:
    patch:
      operationId: updateSalaryRecord
      tags: [admin-finance]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SalaryRecordUpdate" }
      responses:
        "200":
          description: Salary record updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalaryRecordResult" }
        "400": { description: Invalid update }
        "403": { description: Confidential finance data requires the ADMIN role }
        "404": { description: Salary record not found }
        "409": { description: Another salary record already uses this period }
...
  /admin/arrears/{id}:
    patch:
      operationId: updateArrear
      tags: [admin-finance]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArrearUpdate" }
      responses:
        "200":
          description: Arrear updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArrearResult" }
        "400": { description: Invalid update or forbidden state transition }
        "403": { description: Confidential finance data requires the ADMIN role }
        "404": { description: Arrear not found }
        "409": { description: Transfer request is not pending or was already reviewed }
    delete:
      operationId: deleteArrear
      tags: [admin-finance]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: Arrear deleted; only possible before a transfer request exists }
        "400": { description: Invalid arrear id }
        "403": { description: Confidential finance data requires the ADMIN role }
        "404": { description: Arrear not found }
        "409": { description: Records with transfer requests cannot be deleted }
...
components:
  schemas:
...
    Arrear:
      type: object
      required: [id, collaboratorId, periodLabel, amount, currency, status, communicatedReason, transferInstructions, transferRequestedAt, transferRequestStatus, createdAt, updatedAt]
      properties:
        id: { type: integer }
        collaboratorId: { type: integer }
        periodLabel: { type: string }
        amount:
          type: ["string", "null"]
          pattern: '^(0|[1-9][0-9]{0,11})(\.[0-9]{1,2})?$'
          description: Decimal amount is transported as a string to preserve precision; null means legacy amount unknown.
        currency: { type: ["string", "null"], pattern: "^[A-Z]{3}$" }
        status: { type: string, enum: [open, settled, archived] }
        communicatedReason: { type: ["string", "null"] }
        transferInstructions: { type: ["string", "null"] }
        payrollServiceName: { type: string }
        payrollServiceSignature: { type: string }
        transferRequestedAt: { type: ["string", "null"], format: date-time }
        transferRequestStatus: { type: ["string", "null"], enum: [pending, acknowledged, declined, null] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ArrearsList:
      type: object
      required: [arrears]
      properties:
        arrears:
          type: array
          items: { $ref: "#/components/schemas/Arrear" }
    ArrearResult:
      type: object
      required: [arrear]
      properties:
        arrear: { $ref: "#/components/schemas/Arrear" }
    ArrearInput:
      type: object
      additionalProperties: false
      required: [periodLabel, amount, currency, communicatedReason, transferInstructions]
      properties:
        periodLabel: { type: string, minLength: 1, maxLength: 120 }
        amount:
          type: string
          pattern: '^(0|[1-9][0-9]{0,11})(\.[0-9]{1,2})?$'
          description: Decimal amount, sent as a string to avoid floating-point precision loss
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        communicatedReason: { type: string, minLength: 1, maxLength: 5000 }
        transferInstructions: { type: string, minLength: 1, maxLength: 5000 }
        payrollServiceName: { type: string, minLength: 1, maxLength: 120 }
        payrollServiceSignature: { type: string, maxLength: 500 }
        status: { type: string, enum: [open, settled, archived] }
    ArrearUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        periodLabel: { type: string, minLength: 1, maxLength: 120 }
        amount:
          type: ["string", "null"]
          pattern: '^(0|[1-9][0-9]{0,11})(\.[0-9]{1,2})?$'
          description: Decimal string; null explicitly clears the amount to unknown
        currency: { type: ["string", "null"], pattern: "^[A-Z]{3}$" }
        communicatedReason: { type: ["string", "null"], maxLength: 5000 }
        transferInstructions: { type: ["string", "null"], maxLength: 5000 }
        payrollServiceName: { type: string, minLength: 1, maxLength: 120 }
        payrollServiceSignature: { type: string, maxLength: 500 }
        status: { type: string, enum: [open, settled, archived] }
        transferRequestStatus: { type: string, enum: [acknowledged, declined] }
    SalaryRecord:
      type: object
      required: [id, collaboratorId, salaryStatus, periodLabel, communicatedDelayReason, amount, currency, transferInstructions, payrollServiceName, payrollServiceSignature, transferRequestedAt, transferRequestStatus, createdAt, updatedAt]
      properties:
        id: { type: integer }
        collaboratorId: { type: integer }
        salaryStatus:
          type: string
          description: Status as communicated by the administrator
        periodLabel: { type: string }
        communicatedDelayReason: { type: ["string", "null"] }
        amount: { type: ["string", "null"], pattern: '^(0|[1-9][0-9]{0,11})(\.[0-9]{1,2})?$' }
        currency: { type: ["string", "null"], pattern: "^[A-Z]{3}$" }
        transferInstructions: { type: ["string", "null"] }
        payrollServiceName: { type: string }
        payrollServiceSignature: { type: string }
        transferRequestedAt: { type: ["string", "null"], format: date-time }
        transferRequestStatus: { type: ["string", "null"], enum: [pending, acknowledged, declined, null] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SalaryRecordsList:
      type: object
      required: [salaryRecords]
      properties:
        salaryRecords:
          type: array
          items: { $ref: "#/components/schemas/SalaryRecord" }
    SalaryRecordResult:
      type: object
      required: [salaryRecord]
      properties:
        salaryRecord: { $ref: "#/components/schemas/SalaryRecord" }
    SalaryRecordInput:
      type: object
      additionalProperties: false
      required: [periodLabel, salaryStatus, amount, currency]
      properties:
        periodLabel: { type: string, minLength: 1, maxLength: 120 }
        salaryStatus: { type: string, enum: ["Non versé", "Versé"] }
        communicatedDelayReason: { type: ["string", "null"], maxLength: 5000 }
        amount: { type: string, pattern: '^(0|[1-9][0-9]{0,11})(\.[0-9]{1,2})?$' }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        transferInstructions: { type: ["string", "null"], maxLength: 5000 }
        payrollServiceName: { type: string, minLength: 1, maxLength: 120 }
        payrollServiceSignature: { type: string, maxLength: 500 }
    SalaryRecordUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        periodLabel: { type: string, minLength: 1, maxLength: 120 }
        salaryStatus: { type: string, enum: ["Non versé", "Versé"] }
        communicatedDelayReason: { type: ["string", "null"], maxLength: 5000 }
        amount: { type: ["string", "null"], pattern: '^(0|[1-9][0-9]{0,11})(\.[0-9]{1,2})?$' }
        currency: { type: ["string", "null"], pattern: "^[A-Z]{3}$" }
        transferInstructions: { type: ["string", "null"], maxLength: 5000 }
        payrollServiceName: { type: string, minLength: 1, maxLength: 120 }
        payrollServiceSignature: { type: string, maxLength: 500 }
        transferRequestStatus: { type: string, enum: [acknowledged, declined] }
...