# lib/api-zod/src/generated/types/arrearUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArrearUpdateStatus } from './arrearUpdateStatus';
import type { ArrearUpdateTransferRequestStatus } from './arrearUpdateTransferRequestStatus';

export interface ArrearUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  periodLabel?: string;
  /**
     * Decimal string; null explicitly clears the amount to unknown
     * @nullable
     * @pattern ^(0|[1-9][0-9]{0,11})(\.[0-9]{1,2})?$
     */
  amount?: string | null;
  /**
     * @nullable
     * @pattern ^[A-Z]{3}$
     */
  currency?: string | null;
  /**
     * @maxLength 5000
     * @nullable
     */
  communicatedReason?: string | null;
  /**
     * @maxLength 5000
     * @nullable
     */
  transferInstructions?: string | null;
  /**
     * @minLength 1
     * @maxLength 120
     */
  payrollServiceName?: string;
  /** @maxLength 500 */
  payrollServiceSignature?: string;
  status?: ArrearUpdateStatus;
  transferRequestStatus?: ArrearUpdateTransferRequestStatus;
}