# lib/api-zod/src/generated/types/salaryRecordInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SalaryRecordInputSalaryStatus } from './salaryRecordInputSalaryStatus';

export interface SalaryRecordInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  periodLabel: string;
  salaryStatus: SalaryRecordInputSalaryStatus;
  /**
     * @maxLength 5000
     * @nullable
     */
  communicatedDelayReason?: string | null;
  /** @pattern ^(0|[1-9][0-9]{0,11})(\.[0-9]{1,2})?$ */
  amount: string;
  /** @pattern ^[A-Z]{3}$ */
  currency: string;
  /**
     * @maxLength 5000
     * @nullable
     */
  transferInstructions?: string | null;
  /**
     * @minLength 1
     * @maxLength 120
     */
  payrollServiceName?: string;
  /** @maxLength 500 */
  payrollServiceSignature?: string;
}